+++ bin/add_award.py
@@ -51,9 +51,7 @@ def add_award(anthology: Anthology, paper_id: str, title: str) -> Collection:
         log.error(f"Couldn't find paper: {paper_id}")
         return
 
-    award_names = (
-        award.name if isinstance(award, Award) else award for award in paper.awards
-    )
+    award_names = (award.name for award in paper.awards)
     if title.lower() in (name.lower() for name in award_names if name is not None):
         log.warning(f"Award '{title}' already listed for {paper_id}, skipping")
         return